fix: removed balance check from the Shield Context
