+++ ui/contexts/shield/shield-subscription.tsx
@@ -14,11 +14,7 @@ import {
   setShowShieldEntryModalOnce,
   subscriptionsStartPolling,
 } from '../../store/actions';
-import {
-  getSelectedInternalAccount,
-  getUseExternalServices,
-} from '../../selectors';
-import { useAccountTotalFiatBalance } from '../../hooks/useAccountTotalFiatBalance';
+import { getUseExternalServices } from '../../selectors';
 import { selectIsSignedIn } from '../../selectors/identity/authentication';
 import { getIsMetaMaskShieldFeatureEnabled } from '../../../shared/modules/environment';
 import {
@@ -26,7 +22,6 @@ import {
   getIsActiveShieldSubscription,
 } from '../../selectors/subscription';
 import { getIsUnlocked } from '../../ducks/metamask/metamask';
-import { getUserBalanceCategory } from '../../../shared/modules/shield';
 import { useSubscriptionMetrics } from '../../hooks/shield/metrics/useSubscriptionMetrics';
 import { MetaMetricsEventName } from '../../../shared/constants/metametrics';
 
@@ -60,14 +55,8 @@ export const ShieldSubscriptionProvider: React.FC = ({ children }) => {
   const hasShieldEntryModalShownOnce = useSelector(
     getHasShieldEntryModalShownOnce,
   );
-  const selectedAccount = useSelector(getSelectedInternalAccount);
   const { getSubscriptionEligibility: getShieldSubscriptionEligibility } =
     useSubscriptionEligibility(PRODUCT_TYPES.SHIELD);
-  const { totalFiatBalance } = useAccountTotalFiatBalance(
-    selectedAccount,
-    false,
-    true, // use USD conversion rate instead of the current currency
-  );
   const { captureShieldEligibilityCohortEvent } = useSubscriptionMetrics();
 
   /**
@@ -154,29 +143,15 @@ export const ShieldSubscriptionProvider: React.FC = ({ children }) => {
           return;
         }
 
-        if (
-          !selectedAccount ||
-          !isSignedIn ||
-          !isUnlocked ||
-          hasShieldEntryModalShownOnce
-        ) {
+        if (!isSignedIn || !isUnlocked || hasShieldEntryModalShownOnce) {
           return;
         }
 
-        const balanceCategory = totalFiatBalance
-          ? getUserBalanceCategory(Number(totalFiatBalance))
-          : undefined;
-
-        const shieldEligibility = await getShieldSubscriptionEligibility({
-          balanceCategory,
-        });
+        const shieldEligibility = await getShieldSubscriptionEligibility();
 
         if (
           !shieldEligibility?.canSubscribe ||
-          !shieldEligibility.canViewEntryModal ||
-          !shieldEligibility.minBalanceUSD ||
-          !totalFiatBalance ||
-          Number(totalFiatBalance) < shieldEligibility.minBalanceUSD
+          !shieldEligibility.canViewEntryModal
         ) {
           return;
         }
@@ -256,11 +231,9 @@ export const ShieldSubscriptionProvider: React.FC = ({ children }) => {
       isMetaMaskShieldFeatureEnabled,
       isBasicFunctionalityEnabled,
       isShieldSubscriptionActive,
-      selectedAccount,
       isSignedIn,
       isUnlocked,
       hasShieldEntryModalShownOnce,
-      totalFiatBalance,
       getShieldSubscriptionEligibility,
       assignToCohort,
       captureShieldEligibilityCohortEvent,
@@ -271,7 +244,6 @@ export const ShieldSubscriptionProvider: React.FC = ({ children }) => {
     if (
       isMetaMaskShieldFeatureEnabled &&
       isBasicFunctionalityEnabled &&
-      selectedAccount &&
       isSignedIn &&
       isUnlocked
     ) {
@@ -281,7 +253,6 @@ export const ShieldSubscriptionProvider: React.FC = ({ children }) => {
   }, [
     isMetaMaskShieldFeatureEnabled,
     isSignedIn,
-    selectedAccount,
     dispatch,
     isUnlocked,
     isBasicFunctionalityEnabled,
